Cleanup tests, document data contains html and json
