--- tests/data/documents/html-with-toc.js
@@ -1,19 +1,0 @@
-'use strict';
-
-const html = `
-  <html>
-    <head>...</head>
-    <body>
-      <h1>Table of Contents</h1>
-      <h1>Paragraph 1</h1>
-      <h2>Paragraph 1.1</h2>
-      <h1>Paragraph 2</h1>
-      <h2>Paragraph 2.1</h2>
-      <h2>Paragraph 2.2</h2>
-      <h3>Paragraph 2.2.1</h3>
-      <h4></h4>
-    </body>
-  </html>
-`;
-
-module.exports = html;
